Began adding a TextField UI element
Focus is on by default. This will be changed and will be
a part of the Display install where the Display will
constantly check mouse click positions. If it intersects the
UI element, the UI element will "focus" and text input will be
captured. This MUST be moved to a callback so that it captures
key presses and doesn't allow other user key presses to be
captured (which may mean overhauling the key press system
anyways).